Refactor appDemo.ts
