+++ package.json
@@ -20,7 +20,6 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "ioredis": "^5.4.1",
-    "msgpack-lite": "^0.1.26",
     "pg": "^8.11.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -33,8 +32,6 @@
     "@babel/preset-typescript": "^7.24.1",
     "@eslint/js": "^9.2.0",
     "@types/express": "^4.17.21",
-    "@types/lz4": "^0.6.4",
-    "@types/msgpack-lite": "^0.1.11",
     "@types/node": "^20.12.12",
     "@types/pg": "^8.11.6",
     "@types/react": "^18.3.2",
